backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-cloud-firestore==2.17.0
google-auth==2.34.0
passlib[bcrypt]==1.7.4
feedparser==6.0.11
httpx==0.27.2
jinja2==3.1.4
